# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个简单的校园点评网站的后端 API 文档。",
        "title": "杭电点评 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/stores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员直接创建一个已审核通过的店铺。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员 - 店铺"
                ],
                "summary": "创建新店铺",
                "parameters": [
                    {
                        "description": "店铺信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.StoreResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "店铺已存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/admin/stores/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取待审核的店铺列表，支持分页。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员 - 店铺"
                ],
                "summary": "待审核店铺列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.StoreListResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/admin/stores/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员删除指定的店铺（包含关联的评价和图片记录）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员 - 店铺"
                ],
                "summary": "删除店铺",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "404": {
                        "description": "店铺不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/admin/stores/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员批准或拒绝一个店铺。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员 - 店铺"
                ],
                "summary": "更新店铺状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态更新请求 (status: 'approved' 或 'rejected')",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "reason": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的店铺",
                        "schema": {
                            "$ref": "#/definitions/models.Store"
                        }
                    },
                    "400": {
                        "description": "无效的请求",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "店铺不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "店铺状态已被处理，无法再次修改",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
...
        "/stores": {
            "get": {
                "description": "根据多种条件搜索已审核通过的店铺，支持分页、排序和过滤。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "搜索店铺",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词 (名称或地址)",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-created_at",
                        "description": "排序字段 (e.g., -created_at, average_rating)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "店铺分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "店铺状态 (仅管理员可用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.StoreListResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户创建一个新的店铺，需要等待管理员审核。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "创建新店铺",
                "parameters": [
                    {
                        "description": "店铺信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.StoreResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "店铺已存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateStoreRequest": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.StoreResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "average_rating": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "total_reviews": {
                    "type": "integer"
                }
            }
        },
...
        "models.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "auto_created": {
                    "type": "boolean"
                },
                "average_rating": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StoreStatus"
                },
                "total_reviews": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StoreStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "StoreStatusPending",
                "StoreStatusApproved",
                "StoreStatusRejected"
            ]
        },
...
        "problem.Details": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "services.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "services.StoreListResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StoreResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/services.Pagination"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}